Reorganized tests and test components
