+++ package.json
@@ -17,5 +17,6 @@
   "bugs": {
     "url": "https://github.com/oslabs-beta/portara/issues"
   },
-  "homepage": "https://github.com/oslabs-beta/portara#readme"
+  "homepage": "https://github.com/oslabs-beta/portara#readme",
+  "dependencies": {}
 }
