+++ server/package.json
@@ -33,11 +33,11 @@
     "kikstart-graphql-client": "^1.5.0"
   },
   "devDependencies": {
-    "ts-node": "^8.10.1",
-    "@types/jest": "^25.2.1",
-    "jest": "^26.0.1",
+    "@types/jest": "^25.2.3",
+    "jest": "^25.5.1",
     "redis-mock": "^0.49.0",
     "ts-jest": "^25.5.1",
+    "ts-node": "^8.10.1",
     "typescript": "^3.8.3"
   },
   "homepage": "https://github.com/oslabs-beta/portara/blob/staging/README.md"
